make function to delete more one classroom and filter in search

+++ app/Http/Controllers/Grade/GradeController.php
@@ -6,13 +6,14 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\http\Requests\StoreGrades;
 use App\Models\Grade;
+use App\Models\Classroom;
 
 class GradeController extends Controller
 {
     public function store( StoreGrades $request)
     {
-        try {
-
+        try
+        {
             $validated  = $request->validated();
             $Grade      = new Grade();
 
@@ -29,8 +30,8 @@ class GradeController extends Controller
 
     public function update( StoreGrades  $request )
     {
-        try {
-
+        try
+        {
             $validated  = $request->validated();
             $Grades     = Grade::findOrFail($request->id);
 
@@ -49,8 +50,18 @@ class GradeController extends Controller
 
     public function destroy( Request $request )
     {
-        $Grades = Grade::findOrFail($request->id)->delete();
-        toastr()->error(trans('messages.Delete'));
-        return redirect()->route('Grades.index');
+        $MyClass_id = Classroom::where('Grade_id',$request->id)->pluck('Grade_id');
+
+        if($MyClass_id->count() == 0)
+        {
+            $Grades = Grade::findOrFail($request->id)->delete();
+            toastr()->error(trans('messages.Delete'));
+            return redirect()->route('Grades.index');
+        }
+        else
+        {
+            toastr()->error(trans('Grades_trans.delete_Grade_Error'));
+            return redirect()->route('Grades.index');
+        }
     }
 }
